Components have write access to their owner.
Moving towards the idea of having some components contain functionality in
addition to data (and thus being encapsulated).  The 'C struct' style of
components used thus far works well for polling-based systems but not for
event-driven systems.

For example, a transform will be loaded each frame be the rendering system
regardless of whether or not it changed last frame.  On the other hand,
Amber's overlap system only needs to run against entities with altered
transforms.

Even if I change my mind about encapsulated components, there's no harm to
this change.