Engine is a singleton.
Firstly a convenience feature - I don't like having to pass it around
everywhere.  I think this is justifiable given the architecture - an
object should be able to have write access to itself (in the right
circumstances), which currently requires a ref to Engine. The
initialization is a little awkward given the restrictions on calling
bootstrap & getInstance, but it's still pretty hard to screw up.