Scene & Entity cant write to each other.
Write access to the scene shouldn't mean write access to the entities.
This seems counterintuitive (which is why it didn't start this way), but
it produces a much more fine-grained write access (which should reduce
task collision). Write access to the scene is now just used for indexing.

Entities not having write access is less controversial, but it does make
HandleComponent creation & destruction a but funky.  Might be fixed when
all component creation is threaded.