Runs with multiple threads.
Required several changes:

Systems are initialized asynchronously.  This allows for window creation &
OpenGL initialization to occur on a specific (non-main) thread.

Removed the Task parameter capture logic.  Unnecessary when lambdas can
already do this.

Component intialization that makes OpenGL calls is done asynchronously.
Thus, these components have null shaders/textures/meshes for a brief
period - I'd like to refactor this down the road, but I'll leave it for
now since it'll be fine by the next frame.

When adding a task, you can specify which thread to run it on.

Tasks are run in 3 stages: Read -> Write -> Delete.  This will be useful
later when restricting write/delete tasks to eliminate missed writes, but
allowing all reads to run simultaneously.