Editor movement can be undone.
The editted scene is now loaded and maintained separately from what the engine runs on.  Changes to the editted scene are then propogated back to the rendered scene.  This should make it easier to switch (play/pause) down the line and solves the problems with the wrapped component implementaiton.

Includes several other small improvements like namespacing issues.