Style interface directs towards swapping styles
Although swapping individual colors in and out of the only gui style may be more efficient, it was a PITA for the API consumer.  Defining multiple styles up front and making a swap/restore pair of API calls seems nicer.

May be more efficient to store the current style by pointer instead of memcpy'ing, but it's probably not too important.