Data structure copying rework
Don't destroy the old data on copy - leave this up to the user.  This is a c++ idea and doesn't mesh well with the decision to not store element destructors, copy-ers, etc.

Don't require element destruction functions to be re-cast as taking void pointers - this is not strictly legal.  Use macros instead.