Panel priority, mouse blocking, & bug fixes
Panels are now rendered back-to-front, since the first panel MUST be
the top one (detects mouse input first). Added logic for panels to block
mouse events propagating downwards.

When the mouse is pressed inside a panel, it takes priority.  This is
handled by sorting the panels at the end of the frame (or beginning of
the next). Again, the top panel MUST be called first.

New id: focus_id.  This is only used for input boxes, which maintain a
heightened state after the mouse release.  This allows
allow_new_interaction() to check for existing hot and active elements.

Fixed a bug where the panel title bar location was computed before
resizing the panel, which added some resize latency.

Fixed a bug with panel text color being unitialized for default & invis
styles.