Improved hot widget detection; bugfixes
A widget can lose the hot status if it the mouse is covered by a panel
called before it.

A widget can steal the hot status if it is called before the previously
hot widget.

Safeguard hot/active widgets not being called & locking up the
hot/active status.

Fix bug in panel-mouse containment calculation.