Fix shader layout bug
The default layout of vertex attributes on a shader is undefined, and glBindAttribLocation was incorrectly being called after the shader was linked.  This luckily worked on platforms where the attributes happened to be assigned in ascending order, but that behavior is not guaranteed.  Using the layout location specifier in the vertex shader solves this, although it does require that the VBO enum match these layouts.