Remove per-frame gui allocations
Was allocating temporary vertex buffers for circle polygons, polygon type conversions, and polygon decomposition. Instead, store a transient vertex buffer that gets re-used by each algorithm.  Assuming the gui is the same each frame, this buffer will not resize after the first frame.

Had to re-write part of the decomposition algorithm to make use of this.