Basic multi-window support.
Track the number of windows created and change when sdl inits/quits accordingly.
Creates a gl context per window - not really sure if this is the best way, but the documentation is a little lacking.  Seems like a shared context is supported, but I don't have a good reason to try it at this point.  Maybe it improves perf at a greater number of windows?