Temp memory supports variable-sized pages
Can allocate more that 4K bytes in temporary memory.

Unused pgb pages are sorted by size.  When a new page is needed, the
first page big enough for the allocation is used if it exists; otherwise
one is created. All pages have the same number of slots, so larger pages
have a large alignment.

This also removes the dependency on list.h and external memory - all
page header data is stored within the pages themselves.

In theory, it may be better to always put an allocation into the
smallest possible page to avoid wasting memory on alignment.  Haven't
looked into it yet...