Temporary memory forking w/ multiple allocators
A common pattern is for a function to populate an output array from an
input array. If this process requires an intermediate array and the
output array uses the global temporary memory allocator, then the
intermediate array will not be re-captured at function exit.

void func(array(const A) input, array(B) *output)
{
  array(C) temporary = array_create(g_temp_allocator);
  array_append(*output, ...);
  array_destroy(temporary);
}

To solve this, we can create a new allocator for the intermediate data.
Since this allocator can share a heap of pages with other temporary
allocators, it will still re-use existing pages.  However, we can
guarantee that it will not conflict with the output array's allocations
and will be re-captured at function exit.

Note that it may be easier in some situations to have the function
in question initalize the output array itself.  However, this requires
a) passing in the output allocator as well or b) enforcing a documented
allocation scheme.

As always, profile!  This stuff probably isn't necessary for small
arrays, although I don't think it adds too much overhead.