Fix temp memory bug
The pgb allocator wasn't clearing the header info for slots past the restored pointer in pgb_restore(). Consider the following scenario:

A watermark is set on page A.
Page A is used for allocations.
The watermark on page A is restored.
Page A is used up & page B is added.
Page B is popped of due to pgb_free.

To restore the pointer to the last allocation in page A, the allocator will be reading old header info from allocations free'd by the watermark restoration.  This can cause it to use a pointer that is in the middle of a live allocation.