Add color picker widget
Currently requires the color to be stored in floating point format. This
makes it challenging to edit colors currently stored in the gui style.
I'm not yet sure how to rectify this, but for color pickers that are put
into pop-ups, you can simply convert the color format, store & edit it
while the popup is active, then convert back when finished.