Add hint text to (most) widget styles
Improve usability for buttons with pens instead of text.

Whenever widgets are hot, a timer counts down.  If the timer expires,
then the currently hot widget's hint will show, if it exists.

This timer will carry over between widgets, allowing for easy viewing of
multiple hints sequentially.  The timer counts up when no widgets are
hot and resets if a widget becomes active.