Split widget bounds tracking from grid
All render calls extend a bounding box on an ongoing basis.  This is a
bit slower than the old version, but it no longer relies on using a grid
to do so.  It seems worth it to make this feature more useful and
orthogonal to existing features.

This fixes a concern with panels, which can be used without a grid but
would not scroll correctly when doing so.  It pre-empts a similar
concern for scroll areas.