Add scroll area gui widget
A scroll area widget defines a masked region to place other widgets that
may not fit within the region.  The behavior is similar to how panel
scrolling works, but without the need to use a full panel - they can be
used within panels, and *should* even work when nested.

Some work needs to be done to merge the panel's scrolling logic with the
scroll area widget.