gui_drag returns active or triggered by key
Previous version was very unclear: returned a 'changed' bool that was
only true if 1) the widget was active this frame & last frame, or if
2) the widget was triggered & moved by a key press.

Part 2 is fine, but in the case of 1, it's easier to just return true
whenever the widget is active.  The client can add extra logic to handle
a still mouse while dragging - usually the only reason to do so is for
optimization.