Cleanup gui_rect_mcolor()
Directly modifying the vertex color buffer after calling gui_rect is a
bad idea.  It's an un-obvious dependency that makes it prone to future
bugs.  It already had a hidden bug where it would modify unwanted vertex
data, or write before the start of the buffer, if the rectangle was
hidden by the mask.  Using the begin/vert/end system is much better.