Fix TODO for collapsing gui splits at end
Useful for an accordian-style series of panels like the input panels in
TestFit.  If all splits at the end have collapsed panels, these splits
will be positioned at the end of their parent, and the expanded panels
may take up extra room to compensate.

This was implemented by allowing gui splits to have more than 2
children, which are now stored in a linked list.

This is a breaking change, as 1) gui_split2x() can no longer be called
with NULL for the parent, and 2) gui_split2h() and gui_split2v() have
swapped meaning.