Simplify analysis of temp memory for a module
It's hard to analyze how much temp memory a specific module uses, since
the reporting is only for the entire program.  This change gives callers
an easy way to move all the allocated temp memory to a separate heap,
run a body of code, analyze any newly-allocated temp memory, and then
restore the escrowed temp memory.

Note that this requires that the module does not realloc/free anything
that was created in temp memory *before* the module's entry point.