Ensure viewport is correct on frame end
Between a frame begin and frame end, OpenGL drawings may be performed
and the viewport may change, meaning violet has to make sure the
viewport is correct on frame end, before drawing the frame.

This forces us to set the viewport rectangle both on the frame begin and
end, because the screen has to be cleared before external draw calls are
performed to prevent them from being discarded before the window
refresh.