Add error margin for ray intersections
The odds of 2 parallel rays actually having a cross product of 0 seems
near-zero - I usually only see it when the rays are axis-aligned.  I'm
not a huge fan of using an error margin that I don't quite understand,
but checking for parallel has been a major pain in the ass.

This also spills over into segment/line intersections as well.