Partially fix co-planar test for 3d ray intersect
I don't have enough data to find a good error margin for this
calculation, so I'm using error codes to alert the caller so they can
handle it.  I've seen values as high as .00003 for rays that are still
co-planar.  Since this function still sets the t & u interpolation
values even if it thinks the rays aren't co-planar, the caller can
compute the 2 intersection points & check them for equality.