Remove internal error() calls
Doing error handling was a mistake.  It'll take some effort to remove
this from the TestFit codebase, but I can start with the few cases it's
used in violet.

In the out of memory cases, it'll act as if the default error handler
was active, which calls exit().  This may not be the best choice for
other users of the library, but I know that TestFit never checks for
these errors anyways.