Add gui text input flag to select all on focus
Previous behavior was to 1) have a flag to select all on tab-focus and
2) select all when double-clicking.  This was deemed insufficient, so
now text input boxes can be flagged to select all on _any_ type of
focus.

When configured this way, if the user's first click inside the text
field also drags to select some text, we do not override this to select
all.