Allow non-scrollable panels
Try to make panels/scroll areas/grids a little more composable so the
gui can be customized a bit more.

Since this change now requires a panel flag to enable panel scrolling,
it is a breaking API change.

Feels like this could be split up / cleaned up a bit more to make it
easier to use & make the code simpler, but it works for now.