Add assertion for pgb realloc past watermark
If 1) a chunk of memory is allocated from the pgb allocator, 2) a
watermark is saved, and 3) the chunk is realloc'd, then the watermark
will now free the memory.  Since this can cause tough-to-track-down
memory bugs, and we can't think of a situation where it's actually
useful, this change asserts if it encounters that situation in debug
mode.