Fix polyf_contains for point on rectangle diagonal
            x
           *
+---------+
|        *|
|       * |
|      p  |
|         |
+---------+

If PX intersects the polygon exactly at a vertex, we may get an
incorrect number of intersections (0 or 2 instead of 1) due to floating
point error.  This must be very unlikely to happen for us to have made
it this long without hitting this bug.

Using FLT_EPSILON seems like a rather imperfect solution.  This may be
the case that finally makes the error margin transition to violet.