Fix MacOS file dialogs - file extensions
Add an extension dropdown so users don't have to type it out manually.
Fix handling for file patterns with multiple extensions.

Supposedly, MacOS will generate the extension dropdown for you in some
circumstances.  I was not able to discover how to trigger this behavior
- my best guess is it requires use of NSDocument.  Since that seemed
like it would be an intrusive change, I opted for something custom.