Fix triangulation errors around colinear edges
The method used to determine if a point P was on or inside a triangle
ABC wasn't very tolerant of points that were directly on an edge.

A    P    B
 +---+---+
 |
 +
C

We do spend a fair amount of time doing triangulation when rendering a
building, so this is a hot path.  After a few tries I was able to get
the new code to be within 5% of the runtime of the old code.  After
removing a case (commented with a #if 0) that we don't need to support
anymore, this is now about 10% faster than it was.