Fix bizarre profiler crashes on windows
I am very unsure of the root cause of this crash, but it seems that
msvc doesn't like the following syntax (at least for globals):

struct thingy *g_var = &(struct thingy){0};

This compiles, works fine on gcc, and works fine on msvc as long as the
global is NOT thread_local.  The profiler was using this syntax, but
with a thread_local specifier, to avoid an explicit initialization
function and a NULL root block.

I do not know why msvc doesn't like this, but allowing the root block to
be NULL (on initialization & elsewhere) stops the crashes.