Add temp_memory_merge_all()
If we catch a signal, redirect it to a signal handler function, and use
a longjmp to escape a solve, then we lose all pointers to forked temp
allocators, causing their pages to leak.

By storing a global list of all pages created by a heap, we can find
pages that are in use but don't belong to the primary temp allocator and
return them to the heap.