Add ssort() to semi-replace qsort()
There are cases where we don't need a stable sort, but we do need a sort
to be consistent on all platforms.  According to the standard, whether
or not qsort is stable is implementation-defined.  I've found that (my
flavor of) Linux has a stable qsort but Windows does not.

I don't really want to 1) write my own or 2) test someone else's
implementation, so I copied musl libc's implementation of qsort and
exposed it as ssort.