Add OpenGL driver debug logging
Kept behind the usual DEBUG/CHECK_GL #defines so it can be selectively
disabled/enabled for release builds.  I'm not sure if we actually want
to leave this on - it doesn't provide much useful output for my driver,
but I had to implement it in order to find that out.  I'm committing it
with the hope that it helps us track down graphics issues in the future.