Simplify gui widget member styling
It's somewhat common to want to assign a value to a field in each of the
widget states for a single widget type, e.g. setting the font size for
the inactive, hot, and active widget states.  Previously this required
making 3 or 4 separate style push/pop calls for each field, which was
tedious and prone to error.

Since this API got a bit bigger, I also added some documentation.