Add function to change a widget id
I'm not a fan of exposing this at all, but there are a few locations
where TestFit starts a drag operation of a certain type and then
switches it to a different type mid-drag.  An example is splitting a
boundary along an edge.

That behavior isn't new; this change only supports the newest version of
it.  If we can come up with a better way to perform the same action, we
should revert this commit.