Fix text display for high unicode codepoints
The prior implementation assumed that all glyphs in a ttf file were
adjacent unicode codepoints.  This is an incorrect assumption, and it
resulted in invalid memory accesses any time a high codepoint was
requested for rendering.

Note that we must consume more RAM to store a codepoint -> glyph
mapping, more CPU to do the lookup for each rendered codepoint, and more
video memory to store all the glyphs we were previously (and
un-intentionally) skipping over.  I have not benchmarked this change
since the old code was flat out incorrect.