Simplify texture rendering
Due to how we loaded images and font bitmaps, all shaders had to
vertically flip UV coordinates for textures.  This is causing some
confusion in places outside of violet when rendering to textures and
then rendering those textures to the screen.

The downsides are 1) the stb image library has to do extra work when
loading images and 2) we have to set a flag in the stb image library
that isn't thread safe.