Integrate SDL 2.0.18
This requires several hacks to work around new deficiencies in how SDL
handles resizable, borderless windows.  I'm not sure if this is due to
problems with SDL's implementation or if we were relying on un-intended
behavior.  Generally speaking this configuration is a shitshow, and we
should probably look to use a bordered window with custom non-client
drawing instead of what we're doing here.

I chose to proceed with the 2.0.5 -> 2.0.18 update because it fixes some
problems with multi-monitor support on Windows.  There's probably other
bug fixes and improvements, but that's the one we really want.