Undo/Redo: multi-frame mutations
Address the use case when some data mutation might be repeated with a
different payload. If the desired undo behavior is such that it reverts
to the state before any of the repeated actions happened, then instead
of creating a new undo point with every transaction, the previous
transaction's undo point is modified.