Undo/Redo: overhaul high-level state with stores
Some chunk of state is encapsulated as the data of a store_t.
Application state can be composed from multiple store_t, into a
megastore_t. There is a one-to-one correspondence between each
megastore_t and each transaction_system_t. The notion of a store_t as
a polymorphic object reduces the boilerplate to incorporate new app
state with the transaction system.

In other words, app state is encapsulated within a concrete
implementation of store_t