Undo/Redo: change store_kind API
To make the transaction API similar to existing APIs (specifically, gui
style), go with a pop/push approach to setting the store_kind. This way,
we can set and unset the (current) store_kind during a stack frame, and
isolate mutations to a specific store for that stack frame, and allow
higher and lower stack frames to use a different store (all within the
same megastore).