Add cpu_max_sse & os_uname os-dependent functions
The cpu_max_sse function will return an enum indicating the highest
level of SSE supported by the CPU.

The existing cpu_supports_sse41 function now uses cpu_max_sse and
returns true if SSE4.1 or higher is supported.

The os_uname function is a wrapper for uname on Linux and Mac. I have
not fully implemented this for Windows yet (it will just always return
false on Windows).

These are added to support submitting hardware info for analytics.