Add os_device_id to identify unique devices
We want to identify unique devices running TestFit. From what I've read,
Windows, MacOS, and Linux all generate a UUID at installation that can
be used to identify that "installation" that should serve as a fairly
stable way to identify unique devices.

I've implemented this method for Linux only, using the systemd tool
sd_id128_get_machine_app_specific to represent the caller's machine ID
as a UUID in such a way that the original machine ID cannot be
reverse-engineered (AFAIK a cryptographic hash function is used under
the hood to do this).

TODO: implement this for Windows and MacOS