Fix use-after-free crash on failed event execution
We read a value from the event after execution, but if the execution
failed it would already have been destroyed.  This didn't show up on
Linux debug builds, likely because the system allocator didn't
immediately re-use or overwrite the freed memory.  It looks like the
Windows debug allocator does.

The fix was implemented was by Luke, but I committed it since only I
could verify it on Windows.