Don't merge multi-frame events past a timeout
This change keeps the same behavior for rapid changes to the same input
meaning only one undo point is generated, but it makes it so if there
was a pause between changes, it generates separate undo points.
This gives control to the user of the API on what changes should be
considered part of the same multi-frame event, by either triggering an
event every frame to merge, or selectively to create more undo points.

This change was mostly prompted to prevent a specific case where
multiple actions of the same kind were being merged, even though they
were separated by time and by dropping mouse drag. Fixing it this way
means we don't have to worry about assigning some unique id to the
actions, and so we don't have to track mouse state.

Another change required for this to work, is that now the event
timestamp stores the time of the last update it received, not time it
was created.