Undo/Redo: allocate event instance with event_t
In order to get the event base class from an instance of the class
(i.e. concrete event type) we have to make some guarantees about
memory layout. To achieve this, we pack the data for the instance
directly after the metadata. Thus the pair of event_t and its concrete
implementation become a sort of stretchy buffer.

Distinct advantages over the previous, memory disparate solution are
- one malloc call instead of two
- improved cache locality
- any event that implements a destructor must also have a spawner,
  where only dynamic memory allocations, specific to the event, are made