Undo/Redo: allow nested events
We were hitting a wall where many event handlers had to track side-effects
from their function calls (and their sub-calls and so on). This was
quickly leading to behavior that's hard to reason about, easy to get
wrong, and expected to be a nightmare to maintain.

Thus we must soften our previous (self-imposed) requirement that events
cannot spawn other events. Naturally this complicates how the
transaction system executes events and tracks the event history,
but the long-term wins are substantial for API users.

Obvious benefits
- Events need only be concerned with their direct side effects that aren't
  handled in nested events. As such, the Single Responsibility Principal
  is easier to achieve.
- Can have "composite events" that are just dispatchers for several
  other events, together comprising a single undo point

[no-patch]