Undo/Redo: store nested events in a tree
Instead of recording nested events as a linear amalgamation, store them
as members of a tree, where each undo point is the root of the tree, and
all its nested events are children (and their children and so on). This
makes for a natural, clean extension of the system before 831fad35.

[no-patch]