Undo/Redo: disallow merging incompatible events
When a secondary, multi-frame event is followed by its like, and the
subsequent event triggers child events, unless we enforce that the
child events are also multi-frame and of the same kind, there is no good
way to merge into one cumulative event. Instead we elect to prevent a
merge under those conditions.

[no-patch]