Undo/Redo: Add `update_pre` contract callback.
The name of this callback might be subject to change, as it does a poor
job at describing what it's meant to be used for.

The point of `update_pre` is that it's supposed to be similar to
`update`, but called before the execute, not after. Instead of copying
things from the event we've just executed into the first multi-frame
event in order to merge the two, `update_pre` allows you to copy things
from the first multi-frame event into the one we're about to execute.
This means that by calling `update_pre`, we're able to access state only
available at the time of the first multi-frame event for the full
duration of the multi-frame interaction, and allows us to have
accumulation variables that build upon the work already done each frame
and can be used during the execute, as opposed to accumulations
performed during `update` which are only accessible during undo/redo.