Add fixed array utility
This commit adds a way of interfacing with fixed size arrays through the
array interface. This allows callers be able to use all of the utilities
afforded to arrays (e.g. insert / remove / sort), without knowing if the
memory being referenced is a dynamic or fixed size array.

The cap of the fixed array's is initiated as the `countof` its elements.
The fixed array's allocator is set as `g_null_allocator`. which is a new
allocator which always returns `NULL` on `malloc` `realloc` and `calloc`.
This is useful, because it allows the caller of functions like
`array_append` not to have to check if the array needs to grow, or if it
can grow prior to calling the append, by treating the grow on a fixed
array as just a case of "out of memory".
Before this patch, "out of memory" on array functions would always
trigger a fatal error, because this "oom" is much more common and useful
for fixed arrays, the onus of causing a fatal error when out of memory
was moved out of the array, and into the specific allocators where being
out of memory is still supposed to be fatal.

Because the grow operation in fixed arrays is expected to happen, most
functions have been changed to check if the allocations worked, before
reassigning the array, or writing to the result of the allocation.

A lot of code uses the array interface, so the signatures were left
unchanged, but some "_check" variants were added for new code that
may act on fixed arrays.