Revert `fc96932`, remove fixed arrays
Array utilities being used on fixed arrays is dangerous because most
code written to use them assumes a NULL allocation return is fatal. This
means that either all code using array utilities would have to manually
check for "oom", which would be impractical, or a dedicated interface
would have to be used for arrays that may or may not be fixed size.
Because heap-only arrays and fixed+heap arrays would not be
interchangeable, and because it's not possible for the compiler to warn
when someone tries to feed a fixed array to functions that assume the
array can grow, it was decided to remove the concept.