Undo/Redo: tweak algorithm for getting valid undos
It's helpful to be able to query all valid undo/redo targets. To solve
that problem, we introduce a status to each root event to indicate if
we should ever expect it to be undoable/redoable. This is easier to
prove correct than the previous, rather opaque, algorithm for getting
the relevant event_bundle_t, at the expense of increased computational
effort, though we expect that to be minor if not negligible.

[no-patch]