Require serialization for dynamic events
Add save/load functions to the dynamic event factories so that we get
compiler errors if they aren't defined.  While we aren't using the
serialized data quite yet, I want to ensure that new dynamic events have
these functions implemented.

The undo/redo event is dynamic but 1) doesn't use the factories and 2)
can't directly implement the serialization since the system leaves that
up to the application.  I had to stub those methods out to get the code
compiling.

I also added execute() for NOOP events.  I understand why this event
wasn't spawnable & had no contract, but it means that we have to check
for NULL for data that really should never be NULL if we want to handle
all events in the applications event registry.  While it's certainly odd
to spawn a NOOP event, it doesn't seem like a bug.