Add invisible gui input widget
I need an input widget that ignores all mouse input.  This allows us to
fix the polygon drawing tool so that users can click & drag to pan,
click to drop a point, and still type in fixed distances for an edge.

I don't like the fact that you can move the cursor in an invisible input
field - it's unintuitive since you can't see the cursor.  I may try to
remove that, but this commit restores the old (functional) behavior.