Add pgui_panel_open_if_closed utility
For whatever reason, pgui_panel_open and pgui_panel_closed assert that
the panel in question isn't already open or closed (respectively) when
called.

This utility doesn't make assumptions/assertions about the current state
of the panel, but rather is equivalent to pgui_panel_open if the panel
is presently closed when called, and otherwise is simply a no-op.