Add utility to get grid content dimensions
Before this commit, there was no way to get the full width of a vertical
grid, or full height of a horizontal one dynamically without burdening
each strip or widget with the task to grow that dimension.

The new `pgui_grid_content_dimensions` should probably replace
`pgui_grid_dimensions`, but as a precaution, it's left unchanged.