Fix Sentry pgb assertion only in release builds
In `e57e8808`, I added some debugging data to the pgb watermarks to help
track down a tricky memory safety issue.  Since we caught this issue
with an assertion, it was disabled in debug mode.

The recent addition of ASSERT_FALSE_AND_LOG() to start sending tripped
assertions to Sentry changed the behavior such that we started running
one of these checks in release builds even though some of the data isn't
accurate in release builds.  We didn't notice this because 1) it's
pretty uncommon to run release builds locally and 2) we only started
checking assertions in release builds very recently.

Instead of making the assertion debug-only again, I opted to keep the
extra data valid in release builds.  It isn't a large amount of data per
watermark, we don't store many watermarks at once, watermarks are always
placed on the stack, and the operations to update the data are not
intensive.