Clamp panels to be below gui drag area
To prevent input conflicts when a panel is overlapping the window drag
area, this CL updates the panel dimension clamping in Violet so that it
clamps panels to be strictly below any window/gui drag area.

It'd be nice to just use the drag area off the window directly, but
since gui.h doesn't actually know about sdl_gl.h, we have to put the
drag area height on the gui itself and set it from the TestFit app with
gui_set_drag_height when (and if) a custom menu bar is created.

[jira] TF-898