Remove Windows assertion for overlong uuid strings
On other platforms, where we have the uuid library, attempting to parse
an overlong string into a uuid will result in a failure but no debug
assertion.  On Windows, where we have a custom uuid implementation based
on the Win32 APIs, this did trigger an assertion due to a local buffer
being too small.

In order to make the implementation more similar on both platforms, I'm
duplicating the string length check so that we can handle it separately
from errors in the utf8 conversion layer.

[reviewed-by] Matt Gooding